adding support for vs2005 (phase 0)

+++ trunk/Source/C++/Core/Ap4Types.h
@@ -38,11 +38,11 @@
 |   types
 +---------------------------------------------------------------------*/
 typedef int            AP4_Result;
-typedef unsigned long  AP4_Flags;
-typedef unsigned long  AP4_Mask;
-typedef unsigned long  AP4_Cardinal;
-typedef unsigned long  AP4_Ordinal;
-typedef unsigned long  AP4_TimeStamp;
+typedef unsigned int   AP4_Flags;
+typedef unsigned int   AP4_Mask;
+typedef unsigned int   AP4_Cardinal;
+typedef unsigned int   AP4_Ordinal;
+typedef unsigned int   AP4_TimeStamp;
 typedef unsigned long  AP4_Duration;
 typedef int            AP4_Coordinate;
 typedef int            AP4_Distance;
@@ -51,12 +51,11 @@ typedef unsigned int   AP4_UI32;
 typedef unsigned short AP4_UI16;
 typedef unsigned char  AP4_UI08;
 typedef float          AP4_Float;
-//typedef std::string    AP4_String;
 typedef unsigned char  AP4_Byte;
 
 // the rest depends on whether the platform supports 64-bit integers
 #if defined(AP4_CONFIG_HAVE_INT64)
-    // we have 64-bit integers
+    // we have 64-bit integers (not handled yet)
     typedef unsigned long  AP4_Size;
     typedef unsigned long  AP4_Offset;
 #else
